Updated code, still not working

+++ src/main/java/frc/lib/motion/EncoderUtil.java
@@ -3,10 +3,11 @@ package frc.lib.motion;
 public class EncoderUtil {
 
     public static double toDistance(int sensorPosition, int encoderUnitsPerRev, double gearRatio, double wheelDiameter) {
-        return (sensorPosition/encoderUnitsPerRev)*gearRatio*Math.PI*wheelDiameter;
+        return ((double)sensorPosition/(double)encoderUnitsPerRev)*gearRatio*Math.PI*wheelDiameter;
     }
+    
     public static double toVelocity(double velocity, int encoderUnitsPerRev, double gearRatio, double wheelDiameter, double time) {
-        return (velocity/encoderUnitsPerRev)*gearRatio*Math.PI*wheelDiameter;
+        return ((double)velocity/(double)encoderUnitsPerRev)*gearRatio*Math.PI*wheelDiameter;
     }
 
 }
